requirements.txt
alembic==1.8.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bidict==0.22.0
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
click==8.1.3
colorama==0.4.5
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
dnspython==2.2.1
entrypoints==0.4
eventlet==0.33.1
executing==0.8.3
fastjsonschema==2.16.1
Flask==2.2.2
Flask-Cors==3.0.10
greenlet==1.1.2
h11==0.13.0
httptools==0.5.0
idna==2.8
importlib-metadata==4.12.0
importlib-resources==5.8.0
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.1
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pluggy==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
pure-eval==0.2.2
py==1.11.0
pycparser==2.21
pydantic==1.10.2
Pygments==2.12.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==3.13.2
python-socketio==4.3.1
pytz==2022.1
pywin32==304
pywinpty==2.0.6
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
requests==2.22.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.39
sqlparse==0.4.2
stack-data==0.3.0
starlette==0.20.4
terminado==0.15.0
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.25.11
uvicorn==0.18.3
watchfiles==0.17.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.3
Werkzeug==2.2.2
widgetsnbextension==3.6.1
zipp==3.8.1
